Agentic/requirements.txt
# Core dependencies
asyncio>=3.4.3
aiohttp>=3.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI and NLP
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=4.5.0
aioredis>=2.0.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0.0

# MCP and tool integration
mcp>=0.1.0
pydantic-mcp>=0.1.0

# Validation and verification
cerberus>=1.3.0
jsonschema>=4.19.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
cryptography>=41.0.0 